pyproject.toml
[project]
name = "wada-compliance-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.4.2",
    "litellm>=1.72.7",
    "python-dotenv>=1.1.0",
]
